fix bugs related to mutable defaults
Mutable defaults in function definitions ans pydantic model definitons
are parsed once upon application startup and are then re-used for every
instance of the model or call of the function. That means that if a
functions modifies a list that is declared in such a way the
modification will still be present when the function is called again.
This is generally not what is intended